fix: register drain timeout error on the kernel golden surface
Add DaemonDrainTimeoutError to the kernel public runtime golden snapshot,
drop the DaemonDrainOptions/DaemonQueueDrainTarget index exports that had
no non-test consumers, and split the daemon stop drain tests into their
own file to keep daemon-runtime.test.ts within the complexity gate.
