+++ packages/kernel/test/store/daemon-runtime.test.ts
@@ -63,91 +63,6 @@ test("daemon runtime coalesces concurrent evidence reads into one ready repo gen
   });
 });
 
-test("daemon stop drains an in-flight evidence read before it resolves", async () => {
-  await withTempStoreAsync(async (rootDir) => {
-    writeExecutionEvidenceFixture(rootDir, "Shutdown drain");
-    initAuthoredGit(rootDir);
-    commitAuthoredFixture(rootDir);
-    rebuildTaskProjection({ rootDir });
-    let captureStarted!: () => void;
-    let releaseCapture!: () => void;
-    const captureStartedPromise = new Promise<void>((resolve) => {
-      captureStarted = resolve;
-    });
-    const releaseCapturePromise = new Promise<void>((resolve) => {
-      releaseCapture = resolve;
-    });
-    const runtime = createDaemonRuntime({
-      rootDir,
-      materializerPollMs: false,
-      projectionSourceFenceFactory: () => ({
-        capture: async () => {
-          captureStarted();
-          await releaseCapturePromise;
-          return {
-            kind: "stable",
-            identity: "sha256:shutdown-drain",
-            headOid: "shutdown-drain",
-            dirty: false,
-            changedPaths: []
-          };
-        }
-      })
-    });
-    await runtime.start();
-
-    const settled: string[] = [];
-    const read = runtime.queryExecutionEvidencePage({ limit: 1 }).then(
-      () => { settled.push("read"); },
-      () => { settled.push("read"); }
-    );
-    await captureStartedPromise;
-    const stop = runtime.stop().then(() => {
-      settled.push("stop");
-    });
-    await new Promise<void>((resolve) => setImmediate(resolve));
-
-    assert.deepEqual(settled, []);
-    releaseCapture();
-    await Promise.all([read, stop]);
-    assert.deepEqual(settled, ["read", "stop"]);
-  });
-});
-
-test("daemon stop bounds a hung queued write and retains its lock with durable operation tuples", async () => {
-  await withTempStoreAsync(async (rootDir) => {
-    writeExecutionEvidenceFixture(rootDir, "Bounded shutdown");
-    initAuthoredGit(rootDir);
-    commitAuthoredFixture(rootDir);
-    rebuildTaskProjection({ rootDir });
-    const runtime = createDaemonRuntime({
-      rootDir,
-      materializerPollMs: false,
-      interactiveMicroBatchMs: 0,
-      projectionSourceFenceFactory: deterministicProjectionSourceFenceFactory
-    });
-    await runtime.start();
-    void runtime.enqueueBackgroundBatch({ source: "killpoint-hung", run: () => new Promise<never>(() => undefined) });
-    void runtime.enqueueInteractiveWrite({
-      commandId: "recover-after-killpoint",
-      attribution: testAttribution,
-      ops: [docWrite("op-recover-after-killpoint", "task-recover", "note.md", "recover")]
-    });
-    await new Promise<void>((resolve) => setImmediate(resolve));
-
-    await assert.rejects(
-      runtime.stop({ drainTimeoutMs: 25 }),
-      (error: unknown) => error instanceof Error
-        && error.name === "DaemonDrainTimeoutError"
-        && error.message.includes("killpoint-hung")
-        && error.message.includes("recover-after-killpoint")
-        && error.message.includes("op-recover-after-killpoint")
-    );
-    assert.equal(runtime.status().started, false);
-    assert.ok(runtime.status().lockPath, "timeout must retain the write fence instead of releasing the lock");
-  });
-});
-
 test("daemon runtime invalidates only its repo generation after a canonical write", async () => {
   await withTempStoreAsync(async (rootDir) => {
     writeExecutionEvidenceFixture(rootDir, "Canonical invalidation");
